requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astor==0.8.1
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
Django==5.0.14
django-allauth==65.6.0
django-bootstrap3==22.2
django-bootstrap-modal-forms==2.2.0
django-cleanup==9.0.0
django-crispy-forms==2.3
django-forms-bootstrap==3.1.0
django-widget-tweaks==1.4.12
faker==37.1.0
filelock==3.18.0
fire==0.7.0
fonttools==4.57.0
fsspec==2025.3.2
h11==0.14.0
hf-xet==1.1.1
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
imageio==2.37.0
jinja2==3.1.6
joblib==1.5.0
jwt==1.3.1
kiwisolver==1.4.8
lazy-loader==0.4
lmdb==1.6.2
lxml==5.3.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
mssql-django==1.5
networkx==3.4.2
numpy==2.2.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.10.0
paddlepaddle==3.0.0
pillow==11.2.1
pip==25.1.1
pytz==2022.7.1
PyYAML==6.0.2
rapidfuzz==3.13.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==65.5.0
shapely==2.1.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.4.3
stringzilla==3.12.4
sympy==1.14.0
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
